New AgentNodes / extract types
