+++ resources/js/Pages/AgentNodes-basic.tsx
